In our project, there is no certain way because it is a research and development project and new more effective ways can be found during the project. Various methods and techniques will be researched and implemented while the project is in the development process.
In our project, there is no certain way because it is a research and development project and new more effective ways can be found during the project. Various methods and techniques will be researched and implemented while the project is in the development process.In our project, there is no certain way because it is a research and development project and new more effective ways can be found during the project. Various methods and techniques will be researched and implemented while the project is in the development process.In our project, there is no certain way because it is a research and development project and new more effective ways can be found during the project. Various methods and techniques will be researched and implemented while the project is in the development process.In our project, there is no certain way because it is a research and development project and new more effective ways can be found during the project. Various methods and techniques will be researched and implemented while the project is in the development process.In our project, there is no certain way because it is a research and development project and new more effective ways can be found during the project. Various methods and techniques will be researched and implemented while the project is in the development process.In our project, there is no certain way because it is a research and development project and new more effective ways can be found during the project. Various methods and techniques will be researched and implemented while the project is in the development process.